+++ front/src/components/SpotDetail/SpotReview.vue
@@ -8,15 +8,14 @@
             <v-row 
                 class="mx-1"
                 align="center"
-            >
-                <!-- いつかavaterに変えたい -->
-                
+            >   
+                <!-- ユーザーのアバター写真 -->
                 <v-avatar
                     class="mr-5"
                 >
                     <img :src="reviewer_data.user.src">
                 </v-avatar>
-               
+               <!-- ユーザー名（ボタン） -->
                 <v-btn 
                     @click="usernameClickEvent()"
                     text
@@ -25,9 +24,11 @@
                     {{ reviewer_data.user.username }} 
                 </v-btn>
             </v-row>
+            <!-- コメント -->
             <v-row class="mx-5 mt-2">
                 {{ reviewer_data.content.comment }}
             </v-row>
+            <!-- 評価 -->
             <v-row class="mx-3 mt-1">
                 <v-spacer />
                 <v-icon class="mr-1">
@@ -36,35 +37,18 @@
                 {{ reviewer_data.content.score }}
             </v-row>
         </v-container>
-        <user-profile-dialog :showDialog="showUserDialog" :user="reviewer_data.user" @close="closeDialog()"/>
     </v-card>
-    
 </template>
 
 <script>
-import UserProfileDialog from '../share/UserProfileDialog.vue'
 
 export default {
-    components: {
-        UserProfileDialog
-    },
-
     props: {
         reviewer_data: null,
-        // (reviewer_dataの構造) = [{ content: {comment: x , score: y} 
-        //                          { user: {username: a, id: b, src: c} } } ]
-    },
-    data: function() {
-        return {
-            showUserDialog: false,
-        }
     },
     methods:{
-        closeDialog() {
-            this.showUserDialog = false;
-        },
         usernameClickEvent() {
-            this.showUserDialog = true;
+            this.$emit('catchUserData',this.reviewer_data.user)
         },
     },
 }
