+++ front/src/components/UserProfile/SpotListCard.vue
@@ -213,7 +213,7 @@
                 }
             },
             spotInformationPage: function(value) { // spotのカードをクリックしたときに動く関数
-                console.log("spotInformationPage: ", this.spot[value]) // Debug
+                // console.log("spotInformationPage: ", this.spot[value]) // Debug
                 this.showDialog = true;
                 this.selectedSpotID = this.spot[value].spotId;
             },
